+++ sql_scripts/create/payment.sql
@@ -1,6 +1,6 @@
 DROP TABLE IF EXISTS payment;
 CREATE TABLE payment (
-    id_payment INT AUTO_INCREMENT PRIMARY KEY,
+    id_payment BIGINT AUTO_INCREMENT PRIMARY KEY,
     name_payment TINYTEXT NOT NULL,
     number_payment VARCHAR(16) NOT NULL,
     expiry_payment DATE NOT NULL,
